refactor: search bar
